requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0


psycopg2-binary==2.9.9
sqlalchemy==2.0.25
geoalchemy2==0.14.3
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Machine Learning
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
lightgbm==4.3.0
implicit
# API Integrations
# requests==2.31.0
# aiohttp==3.9.1
# google-api-python-client==2.116.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Monitoring & Logging
# loguru==0.7.2

# # Testing
# pytest==7.4.4
# pytest-asyncio==0.23.3
httpx==0.26.0

# Geospatial
shapely==2.0.2
geopy==2.4.1

# Job Scheduling
apscheduler==3.10.4

email-validator==2.1.0
